Better place for the HTML conversion function
that ended up being used nowhere.
